Swapping jackson for cice in the bib merger as well as Jackson sends null when the json contains the value "null"
