+++ sierra_adapter/sierra_bib_merger/src/main/scala/uk/ac/wellcome/platform/sierra_bib_merger/services/SierraBibMergerWorkerService.scala
@@ -3,15 +3,16 @@ package uk.ac.wellcome.platform.sierra_bib_merger.services
 import akka.actor.ActorSystem
 import com.google.inject.Inject
 import grizzled.slf4j.Logging
+import io.circe.generic.extras.auto._
+import io.circe.parser.decode
 import uk.ac.wellcome.metrics.MetricsSender
 import uk.ac.wellcome.models.SierraBibRecord
 import uk.ac.wellcome.models.aws.SQSMessage
 import uk.ac.wellcome.sqs.{SQSReader, SQSReaderGracefulException, SQSWorker}
-import uk.ac.wellcome.utils.JsonUtil
+import uk.ac.wellcome.sierra_adapter.circe._
 
-import scala.concurrent.Future
 import scala.concurrent.ExecutionContext.Implicits.global
-import scala.util.{Failure, Success}
+import scala.concurrent.Future
 
 class SierraBibMergerWorkerService @Inject()(
   reader: SQSReader,
@@ -22,9 +23,9 @@ class SierraBibMergerWorkerService @Inject()(
     with Logging {
 
   override def processMessage(message: SQSMessage): Future[Unit] =
-    JsonUtil.fromJson[SierraBibRecord](message.body) match {
-      case Success(record) => sierraBibMergerUpdaterService.update(record)
-      case Failure(e) =>
+    decode[SierraBibRecord](message.body) match {
+      case Right(record) => sierraBibMergerUpdaterService.update(record)
+      case Left(e) =>
         Future {
           logger.warn(s"Failed processing $message", e)
           throw SQSReaderGracefulException(e)
